Add Welcome, login and sign up screens
