+++ lib/UI_Pages/loginUI_mauryakshitij/signup_screen.dart
@@ -1,21 +1,23 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_ui/screens/signup_screen.dart';
+import 'package:flutter_ui/UI_Pages/loginUI_mauryakshitij/widgets/my_password_field.dart';
+import 'package:flutter_ui/UI_Pages/loginUI_mauryakshitij/widgets/my_text_button.dart';
+import 'package:flutter_ui/UI_Pages/loginUI_mauryakshitij/widgets/my_text_field.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import '../widgets/my_password_field.dart';
-import '../widgets/my_text_button.dart';
-import '../widgets/my_text_field.dart';
 
-class AppLoginScreen extends StatefulWidget {
-  const AppLoginScreen({super.key});
+
+class AppSignUpScreen extends StatefulWidget {
+  const AppSignUpScreen({super.key});
 
   @override
-  State<AppLoginScreen> createState() => _AppLoginScreenState();
+  State<AppSignUpScreen> createState() => _AppSignUpScreenState();
 }
 
-class _AppLoginScreenState extends State<AppLoginScreen> {
+class _AppSignUpScreenState extends State<AppSignUpScreen> {
   final TextEditingController _emailController = TextEditingController();
+  final TextEditingController _usernameController = TextEditingController();
   final TextEditingController _passwordController = TextEditingController();
+  final TextEditingController _confirmController = TextEditingController();
   bool isPasswordVisible = true;
   bool isChecked = false;
 
@@ -41,9 +43,9 @@ class _AppLoginScreenState extends State<AppLoginScreen> {
                   Image(
                       image: isDarkMode
                           ? const AssetImage(
-                              "assets/clueless_logo/logo_dark.png")
+                          "assets/clueless_logo/logo_dark.png")
                           : const AssetImage(
-                              'assets/clueless_logo/logo_light.png'),
+                          'assets/clueless_logo/logo_light.png'),
                       width: width / 6),
                   Column(
                     crossAxisAlignment: CrossAxisAlignment.start,
@@ -63,7 +65,7 @@ class _AppLoginScreenState extends State<AppLoginScreen> {
                           style: GoogleFonts.poppins(
                             fontSize: 18,
                             fontWeight: FontWeight.w400,
-                            color: isDarkMode ? const Color(0xffF0EEEE) : Colors.black,
+                            color: isDarkMode ?const Color(0xffF0EEEE) : Colors.black,
                           ))
                     ],
                   ),
@@ -71,7 +73,16 @@ class _AppLoginScreenState extends State<AppLoginScreen> {
                 ],
               ),
               SizedBox(
-                height: height / 15,
+                height: height / 25,
+              ),
+              SizedBox(
+                width: width / 1.2,
+                child: MyTextField(
+                  labelText: 'Username',
+                  hintText: 'Enter Username',
+                  inputType: TextInputType.text,
+                  controller: _usernameController,
+                ),
               ),
               SizedBox(
                 width: width / 1.2,
@@ -94,77 +105,23 @@ class _AppLoginScreenState extends State<AppLoginScreen> {
                   },
                 ),
               ),
-              Container(
-                padding: EdgeInsets.symmetric(
-                    horizontal: width / 12, vertical: height / 50),
-                alignment: Alignment.centerRight,
-                child: GestureDetector(
-                  onTap: () {},
-                  child: Text(
-                    'Forgot Password?',
-                    style: GoogleFonts.poppins(
-                        color: const Color(0xFF5C5B5B),
-                        decoration: TextDecoration.underline,
-                        fontWeight: FontWeight.w500,
-                        fontSize: 14),
-                  ),
+              SizedBox(
+                width: width / 1.2,
+                child: MyTextField(
+                  labelText: 'Confirm Password',
+                  hintText: 'Enter Password',
+                  inputType: TextInputType.visiblePassword,
+                  controller: _confirmController,
                 ),
               ),
               SizedBox(
-                width: width / 1.1,
-                child: Row(
-                  children: [
-                    Checkbox(
-                      checkColor: Colors.white,
-                      activeColor: Colors.black,
-                      value: isChecked,
-                      onChanged: (bool? value) {
-                        setState(() {
-                          isChecked = value!;
-                        });
-                      },
-                    ),
-                    Text(
-                      'I agree to ',
-                      style: GoogleFonts.poppins(
-                          color: Colors.grey,
-                          fontSize: 10,
-                          fontWeight: FontWeight.w500),
-                    ),
-                    GestureDetector(
-                      onTap: () {},
-                      child: Text(
-                        'Terms and Conditions ',
-                        style: GoogleFonts.poppins(
-                            fontSize: 10, fontWeight: FontWeight.w500),
-                      ),
-                    ),
-                    Text(
-                      'and ',
-                      style: GoogleFonts.poppins(
-                          color: Colors.grey,
-                          fontSize: 10,
-                          fontWeight: FontWeight.w500),
-                    ),
-                    GestureDetector(
-                      onTap: () {},
-                      child: Text(
-                        'Privacy Policy',
-                        style: GoogleFonts.poppins(
-                            fontSize: 10, fontWeight: FontWeight.w500),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              SizedBox(
-                height: height / 40,
+                height: height/40,
               ),
               SizedBox(
                 width: width / 1.2,
                 child: MyTextButton(
                   onTap: () {},
-                  buttonName: 'Login',
+                  buttonName: 'Sign up',
                   bgColor: const Color(0xFF136DD6),
                   textColor: Colors.white,
                 ),
@@ -174,7 +131,7 @@ class _AppLoginScreenState extends State<AppLoginScreen> {
               ),
               Center(
                 child: Text(
-                  'Or login with',
+                  'Or sign up with',
                   style: GoogleFonts.poppins(
                       fontSize: 14, fontWeight: FontWeight.w400),
                 ),
@@ -269,7 +226,7 @@ class _AppLoginScreenState extends State<AppLoginScreen> {
                   mainAxisAlignment: MainAxisAlignment.center,
                   children: [
                     Text(
-                      "Don't have an account? ",
+                      "Already have an account? ",
                       style: GoogleFonts.poppins(
                           fontSize: 14,
                           fontWeight: FontWeight.w400,
@@ -277,13 +234,10 @@ class _AppLoginScreenState extends State<AppLoginScreen> {
                     ),
                     GestureDetector(
                       onTap: () {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute(builder: (context)=>const AppSignUpScreen())
-                        );
+                        Navigator.of(context).pop();
                       },
                       child: Text(
-                        "Sign up",
+                        "Login",
                         style: GoogleFonts.poppins(
                             fontSize: 14,
                             fontWeight: FontWeight.w700,
